+++ tests/loader/flash/test_dataframe_creator.py
@@ -12,84 +12,14 @@ from sed.loader.fel import DataFrameCreator
 package_dir = os.path.dirname(find_spec("sed").origin)
 config_path = os.path.join(package_dir, "../tests/data/loader/flash/config.yaml")
 H5_PATH = "FLASH1_USER3_stream_2_run43878_file1_20230130T153807.1.h5"
-# Define expected channels for each format.
-ELECTRON_CHANNELS = ["dldPosX", "dldPosY", "dldTimeSteps"]
-PULSE_CHANNELS = ["pulserSignAdc", "gmdTunnel"]
-TRAIN_CHANNELS = ["timeStamp", "delayStage", "dldAux"]
-TRAIN_CHANNELS_EXTENDED = [
-    "sampleBias",
-    "tofVoltage",
-    "extractorVoltage",
-    "extractorCurrent",
-    "cryoTemperature",
-    "sampleTemperature",
-    "dldTimeBinSize",
-    "timeStamp",
-    "delayStage",
-]
-INDEX_CHANNELS = ["trainId", "pulseId", "electronId"]
-
-
-def test_get_channels_by_format(config_dataframe):
-    """
-    Test function to verify the 'get_channels' method in FlashLoader class for
-    retrieving channels based on formats and index inclusion.
-    """
-    # Initialize the FlashLoader instance with the given config_file.
-    df = DataFrameCreator(config_dataframe)
-
-    # Call get_channels method with different format options.
-
-    # Request channels for 'per_electron' format using a list.
-    format_electron = df.get_channels(["per_electron"])
-
-    # Request channels for 'per_pulse' format using a string.
-    format_pulse = df.get_channels("per_pulse")
-
-    # Request channels for 'per_train' format without expanding the dldAuxChannels.
-    format_train = df.get_channels("per_train", extend_aux=False)
-
-    # Request channels for 'per_train' format using a list.
-    format_train_extended = df.get_channels(["per_train"])
-
-    # Request channels for 'all' formats using a list.
-    format_all = df.get_channels(["all"])
-
-    # Request index channels only.
-    format_index = df.get_channels(index=True)
-
-    # Request 'per_electron' format and include index channels.
-    format_index_electron = df.get_channels(["per_electron"], index=True)
-
-    # Request 'all' formats and include index channels.
-    format_all_index = df.get_channels(["all"], index=True)
-
-    # Request 'all' formats and include index channels and extend aux channels
-    format_all_index_extend_aux = df.get_channels(["all"], index=True, extend_aux=False)
-
-    # Assert that the obtained channels match the expected channels.
-    assert set(ELECTRON_CHANNELS) == set(format_electron)
-    assert set(TRAIN_CHANNELS_EXTENDED) == set(format_train_extended)
-    assert set(TRAIN_CHANNELS) == set(format_train)
-    assert set(PULSE_CHANNELS) == set(format_pulse)
-    assert set(ELECTRON_CHANNELS + TRAIN_CHANNELS_EXTENDED + PULSE_CHANNELS) == set(format_all)
-    assert set(INDEX_CHANNELS) == set(format_index)
-    assert set(INDEX_CHANNELS + ELECTRON_CHANNELS) == set(format_index_electron)
-    assert set(
-        INDEX_CHANNELS + ELECTRON_CHANNELS + TRAIN_CHANNELS_EXTENDED + PULSE_CHANNELS,
-    ) == set(
-        format_all_index,
-    )
-    assert set(INDEX_CHANNELS + ELECTRON_CHANNELS + PULSE_CHANNELS + TRAIN_CHANNELS) == set(
-        format_all_index_extend_aux,
-    )
+H5_FILE = h5py.File(os.path.join(package_dir, "../tests/data/loader/flash/", H5_PATH), "r")
 
 
 def test_get_index_dataset_key(config_dataframe):
     """Test the creation of the index and dataset keys for a given channel."""
     config = config_dataframe
     channel = "dldPosX"
-    df = DataFrameCreator(config)
+    df = DataFrameCreator(config, H5_FILE)
     index_key, dataset_key = df.get_index_dataset_key(channel)
     group_name = config["channels"][channel]["group_name"]
     assert index_key == group_name + "index"
@@ -104,7 +34,7 @@ def test_get_index_dataset_key(config_dataframe):
 def test_get_dataset_array(config_dataframe, h5_file):
     """Test the creation of a h5py dataset for a given channel."""
 
-    df = DataFrameCreator(config_dataframe)
+    df = DataFrameCreator(config_dataframe, H5_FILE)
     df.h5_file = h5_file
     channel = "dldPosX"
 
@@ -132,7 +62,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
     """Test the method when given an empty dataset."""
 
     channel = "gmdTunnel"
-    df = DataFrameCreator(config_dataframe)
+    df = DataFrameCreator(config_dataframe, H5_FILE)
     df.h5_file = h5_file
     train_id, dset = df.get_dataset_array(channel)
 
@@ -150,7 +80,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
         shape=(train_id.shape[0], 0),
     )
 
-    df = DataFrameCreator(config_dataframe)
+    df = DataFrameCreator(config_dataframe, H5_FILE)
     df.h5_file = h5_file_copy
     train_id, dset_empty = df.get_dataset_array(channel)
 
@@ -194,7 +124,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     """
 #     Test the creation of a pandas DataFrame for a channel of type [per electron].
 #     """
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     df.h5_file = h5_file
 
 #     result_df = df.df_electron
@@ -218,7 +148,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     """
 #     Test the creation of a pandas DataFrame for a channel of type [per pulse].
 #     """
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     train_id, np_array = df.create_numpy_array_per_channel(h5_file, "pulserSignAdc")
 #     result_df = df.create_dataframe_per_pulse(np_array, train_id, "pulserSignAdc")
 
@@ -248,7 +178,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     """
 #     Test the creation of a pandas DataFrame for a channel of type [per train].
 #     """
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     channel = "delayStage"
 
 #     train_id, np_array = df.create_numpy_array_per_channel(h5_file, channel)
@@ -270,7 +200,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     """
 #     Test the creation of a pandas Series or DataFrame for a channel from a given file.
 #     """
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     df.index_per_electron = multiindex_electron
 
 #     result = df.create_dataframe_per_channel(h5_file, channel)
@@ -286,7 +216,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     config = config_dataframe
 #     config["channels"]["dldPosX"]["format"] = "foo"
 
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 
 #     with pytest.raises(ValueError):
 #         df.create_dataframe_per_channel(h5_file, "dldPosX")
@@ -297,7 +227,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     Test the concatenation of channels from an h5py.File into a pandas DataFrame.
 #     """
 
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     # Take channels for different formats as they have differing lengths
 #     # (train_ids can also differ)
 #     print(df.get_channels("all", extend_aux=False))
@@ -328,7 +258,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     config = config_dataframe
 #     config["channels"][channel]["group_name"] = "foo"
 #     index_key = "foo" + "index"
-#     df = DataFrameCreator(config)
+#     df = DataFrameCreator(config, H5_FILE)
 
 #     with pytest.raises(ValueError) as e:
 #         df.concatenate_channels(h5_file)
@@ -341,7 +271,7 @@ def test_empty_get_dataset_array(config_dataframe, h5_file, h5_file_copy):
 #     """
 #     Test the creation of pandas DataFrames for a given file.
 #     """
-#     df = DataFrameCreator(config_dataframe)
+#     df = DataFrameCreator(config_dataframe, H5_FILE)
 #     result_df = df.create_dataframe_per_file(Path(h5_file.filename))
 
 #     # Check that the result_df is a DataFrame and has the correct shape
